Sprint_06/Desafio/Spritn_6/requirements.txt
boto3
python-dotenv
pandas
numpy
matplotlib
